Moving everything out of Main
